Compile fixes for MSYS2 UCRT64
Also added instructions for installing the dependencies on MSYS2.

The selection of the default language is commented out for now. It will
be fixed in another change.

+++ src/gui/npcdialog.cpp
@@ -407,7 +407,7 @@ void NpcDialog::mouseClicked(gcn::MouseEvent &mouseEvent)
     {
         action(gcn::ActionEvent(mNextButton, mNextButton->getActionEventId()));
     }
-    if (mouseEvent.getSource() == mTextBox && isDoubleClick((int)(long)mTextBox))
+    if (mouseEvent.getSource() == mTextBox && isDoubleClick((int)(intptr_t)mTextBox))
     {
         if (mActionState == NPC_ACTION_NEXT || mActionState == NPC_ACTION_CLOSE)
             action(gcn::ActionEvent(mNextButton,
